Fix crash in AppStream PackageKit plugin
The crash on database initialization was caused by an implicit
declaration of the init function in Vala-generated C code, which
resulted in a pointer size which was too small to contain the
DatabaseWrite pointer. Due to some sideeffects, this bug was only
triggered by the PackageKit daemon.
Many thanks to Thiago Macieira for resolving this issue!

Signed-off-by: Matthias Klumpp <matthias@tenstral.net>