Refactor DataProvider code
Years ago, when AppStream was still young, I thought having a
"DataProvider" concept where one could easily plug in metadata parsers
for different formats would be a good idea to faciliate AppStream
adoption.
Today, the DataProvider stuff is useless overhead, which only makes
writing unit-tests a nightmare and adds needless abstraction and
duplication.
Therefore, we remove it entirely with the next version of the AppStream
tools.
This also drops support for the Ubuntu AppInstall format, so people on
Ubuntu will need DEP-11 data soon (which is not a real issue to generate
anymore, thanks to python-dep11).

This change is quite invasive, so expect a few bugs. Also, the removal
of the data-provider stuff made some more simplifications possible
(which I found while refactoring the code), so there will be some
additional changes soon.