as-cache-builder: Cope with no YAML documents
It's possible that a suite/arch will have no packages with AppStream; we
shouldn't crash in that case.

g_strstr_len will return NULL if the string we're searching for ("---")
is not found - this will be the case if there are only the headers and
no further documents. In that case we should read to the end of the
file.