Always build with YAML support
This (and future restructuring of the data serialization and
deserialization code to XML and YAML) will make the code much nicer and
enable some useful new functionality.
Mainly, we will also finally be able to expose YAML-specific routines on
the public API of libappstream now, and allow AppStream to write YAML
documents.