trivial: apt: Only check for cache-writability
The other AppStream directory in /var/lib might not even exist, so the
cache update will never run.
Also, the check for /var/lib is non-fatal in appstreamcli, so we can
omit the additional check.
The /var/cache/app-info directory must be created at package install-
time now (or should we actually check if /var/cache is writable?)