Refactor icon-finding code and fix rare, potential double-free corruption
This is a fun piece... In order to make a memory-corruption / double-
free happen, you need to:

1) Have an AppStream metainfo file with cached icon size information
2) Not have the icon that is referred to in the cache
3) Have no icon search paths set

Since the tmp_icon_path variable was not set to NULL in one occasion, a
second attempt to free it will be made.

Fortunately, the event that all three events above happen has a very low
probability (almost zero, since 3) doesn't happen unless there are other
issues).

Instead of just setting the variable to NULL, I refactored the code so
this kind of mistake doesn't happen again.
The new code will also be faster and export the stock icon property,
which apparently has been forgotten previously.