Be slightly less pedantic about size info when reading screenshot images
This will now only ignore the screenshot if it is a thumbnail without
size information, because that's a strict no in the spec.
We do allow source images without size info now, though. In case such an
image is encountered, width/height of the image are set to 0.

This resolves #34.