xml: Work around non-threadsafe behavior of libxml2
This global error handling sucks, and the workaround is bad too.
But at least you now can have multiple instances of AsMetadata (and even
in multiple threads) without experiencing crashes.
Still, if two instances of AsMetadata are parsing XML in parallel, and
global errors are emitted, it might happen that the wrong AsMetadata
instance receives the error (thanks, libxml2!).
So, no real threadsafety on that, but at least no crashing anymore.