Use a pure PB cache instead of Xapian
This is the first an major step towards deprecating Xapian.
Unfortunately, this massive patch breaks quite a lot of things, but I
didn't want to create an even larger patch that fixes everything.
(Smaller patches are better)
Unfortunately, Xapian is pretty tightly integrated, and replacing it
isn't very easy, but we will manage to do that in the end even without
breaking ABI.