Rename AsDataPool -> AsPool
AsPool is much shorter, and since it's AppStream, there "data" is
redundant and implied.
Since this class is one of the most important ones (it allows access to
the full metadata pool present on the system) having a ahort name makes
sense.
Furthermore, the AsPool can also be used to create search pools of
arbitrary data, if the user of the API doesn't load any cache or
collection metadata.