gir: Correctly annotate ownership transference of GPtrArrays
Apparently, GPtrArray is treated as full container class, which means
that if we set it to automatically free its contents when it is
destroyed, we effectively only transfer the container to the caller and
don't do a full ownership transfer.

This patch fixes all the (wrong) cases where a full transfer is assumed.
This fixes a couple of potential and real crashes happening when using
AppStream via a GIR-aware language such as Python.

Resolves #63