Allow multiple metadata sets in the pool to describe a component
This resolves #57 on the libappstream side.

In order to solve this, we introduce the concept of a "data-id", which
is a unique (but session-specific!) identifier of a metadata set
describing a component.
There can never be two entries with the same data-id, but very well be
two components having the same component-id.
(as opposed to component-ids being the unique thing before)

A data-id has no fixed format (yet; don't rely on any format constraints
when using it!), but libas will make it assume a value like
%{scope}/%{origin_type}/%{appstream_id} where:
scope: The metadata scope, e.g. "system" (from the OS) or "user" (from
the user's home directory) or unset.
origin_type: The origin type, e.g. "limba", "flatpak", "distributor",
...
appstream-id: The known AppStream component ID: