Allow making AppStream consume data from metainfo instead of collection files
This resolves #3

The support for merging in metainfo file data directly when other
components are also available is still a bit rudimentary and can be
extended by updating the merge routine.
In any case, this is a good compromise between the "displayed metadata
should not change when an application is installed versus when it is
just available" vs "I want to see how my metainfo files look like in
software centers" positions.

In order to have AppStream prefer metainfo files over collection data,
set `PreferLocalMetainfoData=true` in appstream.conf