Recognize scaling factor for HiDPI icons
This resolves #21.
We introduce a new "scale" property for icons, setting a scaling factor.
We also scan the icon cache for HiDPI icons if the scale property is set
(the property must be set for this, so we can limit unnecessary search
overhead).

Tools that want to make use of this need to use the get_icons() function
on AsComponent, all previous convenience methods (get_icon_by_size())
ignore the scaling factor for backwards compatibility.
It's likely no convenience methods will be added for this, as it's more
efficient if the client loops over the icon list once to fetch the icons
it needs rather than calling a convenience method multiple times that
will run through the list on each call (adding a convenience methods
would hide this fact while adding little benefit).