trivial: adjust docs to clearly state that licenses are SPDX expressions
Previously the doc was a bit ambiguous about the actual format employed
for licenses. The format expectation is the actual SPDX license expression
format as described in the SPDX specification.
Clearly state this and point to the specification URL for reference on the
format.

Additionally, while SPDX must be parsed as case-insensitive it's considered
best practice to use strict-casing in examples, so change to that.

Closes #122