validator: Make ending dot in component name only a pedantic issue
We should not really make any assumptions on how upstream wants to name
their project. For example, the game "0 A.D." exists, and ending with a
dot is completely justified here.
On the other hand, ending a name with a dor is uncommon and might be a
mistake. So, instead of removing this hint (which I might do anyway at
some point), we demote its priority to only pedantic for now.