docs: <release> tag can have <url> children tags.
This is useful to provide a link specific to a given version of a
software, rather than a generic link meant for the whole project.
Typically you may want to link a release note, or a detailed news which
may give much more information than the <description> and/or probably
more context, images and examples, etc.