spdx: Add some compatibility workarounds for SPDX 3.0
The SPDX team have decided to rename quite a few common tags (with the
'-only' and '-or-later') suffixes and this unfortunately breaks the
world.

Convert the new-style v3 style IDs to the v2 style and thus accept both
forms.

This patch is based on
https://github.com/hughsie/appstream-glib/commit/cdad22c5fdc6579ea6aeccf67372e81372e857b1
of appstream-glib, thanks to Richard Hughes.
It is a follow-up for issue #180 and should address it fully now.