qt: Don't warn about inlining in maintainer mode
In maintainer mode, the warning is an error, causing the build to fail.
Since the -Winline warning is coming from inside Qt (qPrintable), and
does not occur on older Qt versions, there is nothing we can do at the
moment to fix this, so we selectively ignore the warning for the Qt
bindings when in maintainer mode.