apt: Ignore errors post-update
If there are errors we still display them as informational messages so
we know that something bad happened, but we will no longer return an
error code and therefore give users the impression that the whole update
procedure failed.
This should make malformed metadata errors much less dramatic in future.