spec: Clarify format of the appstream: URI
An appstream: URI doesn't have an authority component and is much more
similar to a mailto: URI.
So while adding two slashes will likely work, it's not required or
recommended to have them. Having the two forms listed in one spec page
is really confusing people, so just stick to one form.