cache: Implement various performance optimizations and helpers
Turns out that when adding a large amount of data, most of the time is
spent in g_variant_type_info_query().
This patch turns that down, mostly by not using GVariantDict (not
intuitive, but it works).
Also, a lot of helper functions are implemented, as well as a "floating"
mode for rapidly adding a bunch of components and then persisting them
at once. A limited set of features is available in this mode, just
enough to make AsPool happy.