cache: Allow switch to nosync mode, make insertions threadsafe
This is a massive speed improvement, only getting rid of GVariant had a
similar effect in my tests.
With threadsafe insertions, we could potentially also gain another
performance boost if the single big transaction can be split into
multiple smaller ones (not sure whether that is a good idea yet).