cache: Register addons, remove GVariant
This change registers all addons with their components again and removes
the GVariant usage in the cache and replaces it with dumb string
operations. This change speeds up the cache generation and query time by
180% (!).
In the long run, the string operations will likely also be replaced with
some binary representation.