Refactor component pool to make use of the mmap'ed caches by default
This reduces AppStream's memory usage (outside of building caches) to
almost nothing while allowing improved search features and *much* bigger
amounts of components to be handled at the same time.
This feature will need a lot of testing and improvements in the
following days.

This also makes it possible to get rid of the GVariant-serialized form
of AsComponent, which significantly reduces the maintenance cost of
AppStream. Turns out parsing XML is just as fast and sometimes faster
than the existing GVariant cache.