validator: Implement first bits of tag-based validation
With this new validation system each issue gets a long explanation as
well as a unique short tag which can be used for display it better in
UIs, ignore it or just quickly infer what the problem is.
This is a semi-breaking change as it will change the validator output,
so this will receive further updates in the next few days to make it
work better and display the output better.