Assume releases are sorted in metainfo, fail validation on wrong sorting
Previously, AppStream would read and releases from metainfo files, then
ensure that they were sorted and make a sorted list of releases
available to the user.
Now, we only sort before data is serialized, not when it is loaded.
Additionally, the validator will emit a warning (failing validation)
when releases in a metainfo files are not sorted correctly.