Protect against addon extension loops in the cache
Having those could lead to circular references between components so
that they can not be freed properly anymore, or even worse may crash
AppStream in an infinite recursion when building the data pool.
Since we don't fully control the data we are processing, it's good to
protect against these issues as much as possible, even though they
should not ever happen on a system that receives well processed
metadata.