Interleave translated/untranslated paras when generating metainfo XML
This resolves #246

Previously, we did not alternate localized/unlocalized description
paragraphs when generating metainfo XML for simplicity.
Now we try to approximate what a human or gettext would write when
generating the output.
All tests pass with this code, but more testing in the real world is
welcome to rule out that this hits some odd corner cases.