content-rating: Add content rating system APIs from gnome-software
These functions were added to libappstream-glib and were previously
requested for libappstream as well, so we are including them here as
well since libappstream and libappstream-glib can not be used together
in the same application.

The content rating system translations were originally written for
gnome-software and this patch was originally introduced by Philip
Withnall.

This code was originally written by:
 • Richard Hughes <richard@hughsie.com>
 • Piotr Drąg <piotrdrag@gmail.com>
 • Philip Withnall <withnall@endlessm.com>
 • Jordi Mas <jmas@softcatala.org>

It was previously in `src/gs-content-rating.[ch]` in gnome-software,
under the GPL-2.0+; and in `libmalcontent-ui/gs-content-rating.[ch]` in
malcontent, also under the GPL-2.0+. It has been relicensed from
GPL-2.0+ to LGPL-2.1+ for the move to appstream-glib, and all the
previous contributors have agreed. See
https://github.com/hughsie/appstream-glib/pull/364#issuecomment-632773905
and below for their agreements.

In moving it from one project to the other, I have added documentation
comments and tweaked some of the APIs to make them more usable.

Signed-off-by: Philip Withnall <withnall@endlessm.com>
Signed-off-by: Matthias Klumpp <matthias@tenstral.net>