Revamp version comparison algorithm
The new variant is derived from the APT/dpkg version comparison
algorithm by Julian Andres Klode and has been adjusted slightly to pass
not only for Debian versions but also versions in the AppStream
testsuite. It should be applicable for most cases where version numbers
need to be compared, and is also quite fast.

Resolves: #288