trivial: Properly mark as-profile as internal, but exported API
This is just cosmetic. AsProfile is considered internal API but is
exported so other modules of the AppStream project, like appstreamcli
and libappstream-compose can use it. It is not intended to be used
outside of the AppStream project.
This change just reflects that in the filelist and formalizes it - the
profiles header was never exported properly before and could not have
been used by external projects.