Switch back from Soup to cURL for HTTP(S)
It turns out that people do prefer cURL over libsoup for
minimal/embedded installations, and for some odd reason *do* want
libappstream on those...
Since our use of libsoup is quite light and libappstream is supposed to
not have a lot of heavy dependencies, we just switch back to cURL as
used before. This time though, it's wrapped in a new internal AsCurl
class for convenience.