search: Make search a bit more strict
We now try to match all search terms, unless that doesn't yield any
results, in which case we try to be a bit more forgiving and also allow
components which match fewer terms to be returned.
Overall though, search results should be much more specific now, but you
should still be able to find things without knowing exactly what you are
looking for.
Resolves: #307