Share one user-owned read-only system metadata cache between all applications
This will conditionally create a new user-owned cache for system-wide
metadata, if none exists in the system's cache directories yet. This
cache is read-only and shared between all running applications, which
should dynamically recreate it in case it runs out of date.

Clients can configure this behavior by setting/unsetting the
AS_CACHE_FLAG_REFRESH_SYSTEM flag on an AsPool.

Resolves: #306