search: Perform partial token matches instead of prefix matches
If exact matches yielded no results, we now perform a partial match
instead of a prefix match - surprisingly, this is even in the worst case
not that much worse than a prefix match, and the compiler does an
excellent job at optimizing these string comparisons.
Also, we don't cheat and remove search terms in case of no matches
anymore.