component: Don't strip ";" from keywords before translating them
We're failing to translate strings now, when the keywords ends with a
";" as they almost always do, e.g.:

  Keywords=Drivers;Repositories;Repository;PPA;

because we strip the trailing semicolon off before translating, in order
to avoid having empty elements in the output.

Instead what we can do is process the list passed to
`as_component_set_keywords` and remove empties at that point. This
allows us to not modify the string retrieved from the desktop file, and
fixes finding translations.