Work around odd heap corruption issues when using threads+libxml2
This *still* doesn't fully address the issue - if we also use the
validator (which also parses XML) in parallel to parsing XML to get the
component, we still get the heap corruption issue. If we only parse XML,
the issue does not appear. Mutex'ing these functions fixes the issue.
Adding a mutex around the data processing functions does not fix the
issue.
The issue is only ever gone if every single thing touching an object
associated with an active xmlDoc has been mutex'ed, which seems very
excessive. It would also be weird if an application can only ever have
one xmlDoc open at the same time.
This issue will need some further research, but the current change
addresses a lot of issues already, so we can have it as band-aid.