Add assertion against as_get_resource() returning NULL
Sometimes in multithreaded applications, as_get_resource() (and
therefore g_static_resource_get_resource) returns NULL. These functions
is called *a lot*, and this effect is very rare, but we currently crash
with a segfault once that happens.
Add an assertion for now to better catch this error, but we do need to
figure out what causes it in the long run. The GLib code looked fine to
me so far.