ci: Add Fedora to the CI environment (#347)
As Fedora ships updated GNOME components often earlier than anyone else,
it makes sense to test this to ensure that this continues to work.

DAPS is not broadly packaged, so the documentation cannot be built
on most distributions. In particular, it's not available on Fedora
right now, so disable it to make the build succeed.