spec: Add support for source locales
This allows specifying the locale of the source translatable strings
in a component. The vast majority of the time, this is `en_US`, but
that’s not guaranteed. Add a `source_locale` attribute to allow changing
this default.

The attribute has been added to the `<translation/>` element so that it
can be used by the metadata generator to synthesize a `<lang/>` element
for the source locale.

It’s not been added to the `<languages>` element, as it’s assumed that
element is auto-generated and hence can include the source locale
explicitly as a `<lang/>` element.

Signed-off-by: Philip Withnall <pwithnall@endlessos.org>

Helps: #345