Emit URLs to choosealicense.com for certain popular licenses
This should be a nicer representation for users when they click on a
license to see what the license is about.
I would like a repository like this to be owned by a more neutral party
than Microsoft, but the info on that page currently is great and well
curated, so we might as well use it for the time being.