Add manually curated list of DFSG-free licenses
Resolves: #357
This is a bit of a workaround, but I do not want AppStream to become an
authority on license-freeness. The DFSG/FSF/OSI are the better authority
on that. But with the DFSG override list we can at least say "if it's
free enough for Debian main, it's considered to be a free license by
AppStream too".