ascli: news-to-metainfo: Allow only marking the last entries as translatable
This is useful for projects which only want the latest X entries to be
marked for translation, and ignore any older, historical entries as to
not cause needless work for translators.

This change required a bit of refactoring, which caused this commit to
become larger than I wanted it to be.

Resolves: #350