compose: Process video screenshots as well
To inspect a video, we need FFmpeg's `ffprobe` binary currently. If this
binary is not found, we will just download the video file and blindly
trust the human metadata author to have filed out everything correctly.

The file size limit feature was not yet ported from appstream-generator,
and disabling just screenshot processing is also not yet possible.