Improve FTS search performance
We now take at least some of the match score into account again (to get
them all, libxmlb needs new features, unfortunately). We also will not
load the token data from cache, as it is generally not needed to be in
memory post-search (saves a few milliseconds in total on searches).