Add an auto-reload mechanism to refresh cache sections when data changes
If one file changes we will have to reload the respective section, as
the change may impact other components as well (due to the presence of
merge components, or simply because it may replace or conflict with an
existing component). However, since the cache is split into sections,
reloading is a lot cheaper now, and it also transparently done in a
background thread.
This feature is opt-in at the moment, so API users will need to add the
AS_POOL_FLAG_MONITOR to their pool flags.

Resolves: #337