cache: Protect against DOM with missing values
When AppStream is serializing data internally, it may creates nodes with
no value. When the final XML is generated, those will be ignored, so we
need to ignore them as well when transmogrifying into XbBuilderNodes.
This patch also adds some additional validation for text nodes, to
improve correctness of the translation (otherwise some nodes may end up
with mixed states).
This will likely need further refinement in order to make the
<description/> tag work properly.