Fix broken markup by using itstool to generate final translated metainfo
msgfmt escapes all XML, even though we don't want that for MetaInfo
files description markup. Itstool does not have this issue, so let's use
that instead. This feature should probably be part of the i18n Meson
module in the future, so it is less of a hack to use.