Improve search by combining xmlb's native node tokens with tokens-as-tags
AppStream generates a lot of tokens to search through, more than libxmlb
can currently handle. So we split those entries with a lot of tokens
into their own subtree and search them individually as nodes with a
different priority, while using the xmlb-native tokens to search for the
fewer, but usually much higher value tokens as present in for example
name and summary tags.
This approach yields not only much better search results, but also
speeds up cache generation and search by about 20msec for search and
~150msec for cache generation.