compose: Port over font support from appstream-generator
This is the C equivalent to what appstream-generator does for font
handling.
This is the last handler that needed to be ported, so it is likely that
we can now seal the compose API and only expose the necessary things.
The appstream-generator can then use the high-level compose code and
will no longer need any special treatment.