qt: Always use C library for enum-to-string conversions
The Qt way is of course nicer, but it's yet another place to keep a
string table up to date, so calling the C library for that conversion is
a lot more convenient in the long run and easier to maintain. And unless
people are doing insane things with the ASQt API, the additional
overhead should not be noticeable.

CC: #380