Relicense remaining GPL-2.0 code to LGPL-2.1+
It makes sense for AppStream to be under one consistent license, so we
can freely move code from the tools into the libraries and vice versa,
without having to ask every author for permission first.
The GPL also has almost no advantages for the tiny amount of code in
AppStream that fell under it, our CLI tools may just as well be
LGPL-licensed.