meson: Bump minimum version to 0.62
This is actually a rather steep requirement that I am not totally
comfortable with, as it complicates backports. However, we have
accumulated a really huge amount of workarounds for Meson issues, and
there are subtle bugs in older versions which do cause occasional
trouble. All of this is fixed in 0.62, so to lessen the maintenance
burden on this stuff, we just depend on that version with this release.