compose: Don't create bad data when localized screenshots exist
This helps #395, but the actual proper fix is to implement handling of
screenshot localization in compose properly, as the current code is
completely oblivious to that.
In addition to that change, language compatibility checks for the C
locales also don't seem to work as expected on Alpine, which needs
further investigation as well.