compose: Allow building without SVG support
Apparently, bootstrapping a system with Rust in the toolchain is still a
major pain, and since rsvg is being rewritten in Rust, AppStream is
dragging it in if compose is enabled.
Of course, building AppStream without SVG support is not a great idea,
unless the only thing you are using appstream-compose for is non-GUI
applications.
As this is apparently the case for some Flatpak runtimes, we allow
disabling rsvg, but also complain very loudly as soon as we encounter an
SVG icon. This change should simplify early bootstrap and allow building
the complete appstream-compose at a later time in the build process.

Resolves: #434