Add a hack to clarify proper PtrArray element ownership for language bindings
In the GPtrArrays that many of our functions return, the elements
contained within are owned by the GPtrArray and are dereferenced as soon
as the container itself is free'd. So, a "transfer full" annotation
would make sense, but in that case implementations like the one in
Python will try to free the elements within, leading to an obvious
double-free. If we use a "transfer container" annotation, Python works,
but then Vala breaks as it now assumes weak references of the contained
elements.
There is no winning this, and it can produce quite subtle bugs. If we
*don't* use the container-owns-elements feature of GPtrArray, we
increase the chance of bugs in C code and make the C code quite a bit
more verbose.

According to
https://gitlab.gnome.org/GNOME/gobject-introspection/-/issues/305#note_1010673
this issue will never get addressed upstream, so the only somewhat sane
thing we are left with here is to add dedicated helper functions
exclusively for languages using bindings, so they all can do the right
thing.
Unfortunately, Vala's vapigen doesn't seem to honor "rename-to"
directives, so we need to both add a "skip" on the C function *and* need
to rename the previous function explicitly for Vala.
Python & Co. seem to do the right thing here though.