ci: Ignore failures on Fedora Rawhide
This test fails extremely frequently, not due to faults of our own but
due to packaged Git snapshots of GCC or various packaging issues.

Right now we are again hit by a GCC bug in its development version, so
we ignore failures this check for now to keep the CI useful.