Make AppStream use BCP47 for locale in XML data
While investigating why zh_TW/zh_CN translations were not showing up
in AppStream-based software centers, we found out that this was due to
the locale being listed as "zh-TW" in the XML. This was first noticed
at KDE, which was going to edit their tools to use POSIX locale in the
XML instead[1].

When using XML, the contents of the xml:lang tag are not arbitrary or
UNIX locale though, but need to follow the IETF BCP47 specification.
So by assuming POSIX locale, AppStream was doing the wrong thing for a
really long time.

Since this change breaks compatibility, I initially just wanted to keep
the "POSIX locale" way, because it is also just a lot simpler. But a
very large amount of projects (roughly 40%) do generate the data the
BCP47 way, and adjusting each and every tool to do it "wrong" while
there is a specification for it is an extremely tough call.

So, we treat this as a bug in AppStream and resolve it here. The "POSIX
in xml:lang" behavior is common because many projects invoke msgfmt in a
particular way, so resolving this issue in GNU Gettext is also needed
(and a patch is being prepared for review).

This patch does take care to be as backwards compatible to the previous
code as reasonable, so many projects should not notice any difference to
before, and only specialized projects like appstream-generator may need
to adjust.

See
https://lists.freedesktop.org/archives/appstream/2023-April/000401.html
for a detailed explanation.