Ensure stemmer always has the right locale and token-search works
This resolves #507
The problem was two-fold: First, for some locale stemming was loaded
wrong, so the wrong word stems were generated. And second, due to a bug
in libxmlb, we never actually searched for tokens, which likely slowed
down search quite a bit and made its results much worse.

This patch implements a workaround for the libxmlb bug and fixes the
AppStream bug. In future, we will likely need to bump the libxmlb
dependency instead to avoid the ugly workaround.